chore(tests): consolidate tests under top-level tests/ directory
Move all test files from per-module test/ subdirectories into a single
tests/ tree that mirrors the source layout. Update relative imports,
the npm test glob, .hlxignore, and TESTING.md references accordingly.
